Migrate to new dependencies
