+++ buildSrc/src/main/kotlin/BuildExtensions.kt
@@ -31,10 +31,10 @@ import io.spine.internal.dependency.GradleDoctor
 import io.spine.internal.dependency.Kotest
 import io.spine.internal.dependency.Kover
 import io.spine.internal.dependency.Ksp
-import io.spine.internal.dependency.McJava
-import io.spine.internal.dependency.ProtoData
-import io.spine.internal.dependency.ProtoTap
 import io.spine.internal.dependency.Protobuf
+import io.spine.internal.dependency.spine.McJava
+import io.spine.internal.dependency.spine.ProtoData
+import io.spine.internal.dependency.spine.ProtoTap
 import io.spine.internal.gradle.standardToSpineSdk
 import org.gradle.api.Project
 import org.gradle.api.Task
