+++ buildSrc/src/main/kotlin/DependencyResolution.kt
@@ -50,8 +50,8 @@ import io.spine.internal.dependency.OpenTest4J
 import io.spine.internal.dependency.Plexus
 import io.spine.internal.dependency.Protobuf
 import io.spine.internal.dependency.Slf4J
-import io.spine.internal.dependency.Spine
 import io.spine.internal.dependency.Truth
+import io.spine.internal.dependency.spine.Spine
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
@@ -195,9 +195,10 @@ fun Project.forceSpineBase() {
  * Forces configurations containing `"proto"` in their names (disregarding the case) to
  * use [Spine.baseForBuildScript].
  */
+@Suppress("unused")
 fun Project.forceBaseInProtoTasks() {
     configurations.configureEach {
-        if (name.toLowerCase().contains("proto")) {
+        if (name.lowercased().contains("proto")) {
             resolutionStrategy {
                 force(Spine.baseForBuildScript)
             }
