+++ buildSrc/src/main/kotlin/DokkaExts.kt
@@ -35,7 +35,6 @@ import org.gradle.api.file.FileCollection
 import org.gradle.api.tasks.TaskContainer
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.bundling.Jar
-import org.gradle.configurationcache.extensions.capitalized
 import org.gradle.kotlin.dsl.DependencyHandlerScope
 import org.jetbrains.dokka.DokkaConfiguration
 import org.jetbrains.dokka.base.DokkaBase
@@ -70,8 +69,10 @@ fun DependencyHandlerScope.useDokkaWithSpineExtensions() {
 private fun DependencyHandler.dokkaPlugin(dependencyNotation: Any): Dependency? =
     add("dokkaPlugin", dependencyNotation)
 
-private fun Project.dokkaOutput(language: String): File =
-    buildDir.resolve("docs/dokka${language.capitalized()}")
+private fun Project.dokkaOutput(language: String): File {
+    val lng = language.titleCaseFirstChar()
+    return layout.buildDirectory.dir("docs/dokka$lng").get().asFile
+}
 
 fun Project.dokkaConfigFile(file: String): File {
     val dokkaConfDir = project.rootDir.resolve("buildSrc/src/main/resources/dokka")
@@ -183,7 +184,7 @@ fun Project.dokkaKotlinJar(): TaskProvider<Jar> = tasks.getOrCreate("dokkaKotlin
  * The task `"publishToMavenLocal"` is excluded from the check because it is a part of
  * the local testing workflow.
  */
-fun DokkaTask.isInPublishingGraph(): Boolean =
+fun AbstractDokkaTask.isInPublishingGraph(): Boolean =
     project.gradle.taskGraph.allTasks.any {
         with(it.name) {
             startsWith("publish") && !startsWith("publishToMavenLocal")
@@ -216,7 +217,7 @@ fun Project.dokkaJavaJar(): TaskProvider<Jar> = tasks.getOrCreate("dokkaJavaJar"
 fun Project.disableDocumentationTasks() {
     gradle.taskGraph.whenReady {
         tasks.forEach { task ->
-            val lowercaseName = task.name.toLowerCase()
+            val lowercaseName = task.name.lowercased()
             if (lowercaseName.contains("dokka") || lowercaseName.contains("javadoc")) {
                 task.enabled = false
             }
