+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/spine/McJava.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.internal.dependency
+package io.spine.internal.dependency.spine
 
 /**
  * Dependencies on Spine Model Compiler for Java.
@@ -33,26 +33,47 @@ package io.spine.internal.dependency
  */
 @Suppress(
     "MemberVisibilityCanBePrivate" /* `pluginLib()` is used by subprojects. */,
-    "ConstPropertyName"
+    "ConstPropertyName",
+    "unused"
 )
 object McJava {
     const val group = Spine.toolsGroup
 
-    /** The version used to in the build classpath. */
+    /**
+     * The version used to in the build classpath.
+     */
     const val dogfoodingVersion = "2.0.0-SNAPSHOT.244"
 
-    /** The version to be used for integration tests. */
+    /**
+     * The version to be used for integration tests.
+     */
     const val version = "2.0.0-SNAPSHOT.247"
 
+    /**
+     * The ID of the Gradle plugin.
+     */
     const val pluginId = "io.spine.mc-java"
 
+    /**
+     * The library with the [dogfoodingVersion].
+     */
     val pluginLib = pluginLib(dogfoodingVersion)
+
+    /**
+     * The library with the given [version].
+     */
     fun pluginLib(version: String): String = "$group:spine-mc-java-plugins:$version:all"
 
     /** The artifact reference for forcing in configurations. */
-    @Suppress("unused")
     const val pluginsArtifact: String = "$group:spine-mc-java-plugins:$version"
 
+    /**
+     * The `mc-java-base` artifact with the [version].
+     */
     val base = base(version)
+
+    /**
+     * The `mc-java-base` artifact with the given [version].
+     */
     fun base(version: String): String = "$group:spine-mc-java-base:$version"
 }
