+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/publish/Publications.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -174,14 +174,14 @@ internal class StandardJavaPublicationHandler(
      *
      * A typical Maven publication contains:
      *
-     *  1. Jar archives. For example: compilation output, sources, javadoc, etc.
-     *  2. Maven metadata file that has ".pom" extension.
-     *  3. Gradle's metadata file that has ".module" extension.
+     *  1. Jar archives. For example, compilation output, sources, javadoc, etc.
+     *  2. Maven metadata file that has the ".pom" extension.
+     *  3. Gradle's metadata file that has the ".module" extension.
      *
-     *  Metadata files contain information about a publication itself, its artifacts and their
+     *  Metadata files contain information about a publication itself, its artifacts, and their
      *  dependencies. Presence of ".pom" file is mandatory for publication to be consumed by
      *  `mvn` build tool itself or other build tools that understand Maven notation (Gradle, Ivy).
-     *  Presence of ".module" is optional, but useful when a publication is consumed by Gradle.
+     *  The presence of ".module" is optional, but useful when a publication is consumed by Gradle.
      *
      * @see <a href="https://maven.apache.org/pom.html">Maven – POM Reference</a>
      * @see <a href="https://docs.gradle.org/current/userguide/publishing_gradle_module_metadata.html">
