+++ buildSrc/src/main/kotlin/module.gradle.kts
@@ -30,9 +30,12 @@ import io.spine.internal.dependency.ErrorProne
 import io.spine.internal.dependency.JUnit
 import io.spine.internal.dependency.Jackson
 import io.spine.internal.dependency.Protobuf
-import io.spine.internal.dependency.Spine
 import io.spine.internal.dependency.Truth
-import io.spine.internal.dependency.Validation
+import io.spine.internal.dependency.spine.CoreJava
+import io.spine.internal.dependency.spine.Logging
+import io.spine.internal.dependency.spine.Spine
+import io.spine.internal.dependency.spine.ToolBase
+import io.spine.internal.dependency.spine.Validation
 import io.spine.internal.gradle.javac.configureErrorProne
 import io.spine.internal.gradle.javac.configureJavac
 import io.spine.internal.gradle.javadoc.JavadocConfig
@@ -152,10 +155,10 @@ fun Module.forceConfigurations() {
                     Spine.base,
                     Spine.time,
                     Spine.testlib,
-                    Spine.toolBase,
-                    Spine.Logging.libJvm,
-                    Spine.Logging.middleware,
-                    Spine.server,
+                    ToolBase.lib,
+                    Logging.libJvm,
+                    Logging.middleware,
+                    CoreJava.server,
                     "io.spine.validation:spine-validation-java-runtime:$validationVersion",
 
                     Jackson.core,
