+++ proto/configuration/build.gradle.kts
@@ -24,8 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.ProtoData
-import io.spine.internal.dependency.Spine
+import io.spine.internal.dependency.spine.ProtoData
 
 plugins {
     `build-proto-model`
@@ -33,6 +32,5 @@ plugins {
 
 dependencies {
     api(ProtoData.api)
-    implementation(Spine.base)
     implementation(project(":java-runtime"))
 }
